[Feat] #108 - 족보에 식당 추가 API 연동

+++ Hankkijogbo/Hankkijogbo/Network/Zip/Request/PostHankkiToZipRequestDTO.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct PostHankkiToZipRequestDTO: Codable {
-    let favoriteId: Double
-    let storeId: Double
+    let favoriteId: Int64
+    let storeId: Int64
 }
